+++ src/BulletCollision/CollisionShapes/btMiniSDF.cpp
@@ -114,13 +114,13 @@ bool btMiniSDF::load(const char* data, int size)
 	unsigned long long int n_cells0;
 	ds.read(n_cells0);
 	m_cells.resize(n_cells0);
-	for (int i = 0; i < n_cells0; i++)
+	for (unsigned long long i = 0; i < n_cells0; i++)
 	{
 		unsigned long long int n_cells1;
 		btAlignedObjectArray<btCell32>& cells = m_cells[i];
 		ds.read(n_cells1);
 		cells.resize(n_cells1);
-		for (int j = 0; j < n_cells1; j++)
+		for (unsigned long long j = 0; j < n_cells1; j++)
 		{
 			btCell32& cell = cells[j];
 			ds.read(cell);
@@ -132,13 +132,13 @@ bool btMiniSDF::load(const char* data, int size)
 		ds.read(n_cell_maps0);
 
 		m_cell_map.resize(n_cell_maps0);
-		for (int i = 0; i < n_cell_maps0; i++)
+		for (unsigned long long i = 0; i < n_cell_maps0; i++)
 		{
 			unsigned long long int n_cell_maps1;
 			btAlignedObjectArray<unsigned int>& cell_maps = m_cell_map[i];
 			ds.read(n_cell_maps1);
 			cell_maps.resize(n_cell_maps1);
-			for (int j = 0; j < n_cell_maps1; j++)
+			for (unsigned long long j = 0; j < n_cell_maps1; j++)
 			{
 				unsigned int& cell_map = cell_maps[j];
 				ds.read(cell_map);
